Refactor to split out many features
Many of the features in ONE had no reason to be ONE specific, so they've
been pulled out and made available as AnyEvent::* modules.  There are
still ONE loaders for them so you can easily do things like:

    use ONE qw( Collect Sleep Timer );

We also add ONE::Capture which uses AnyEvent::Capture to make it easy
to call an event synchronously.